Don't use pristine-tar, fake upstream tag.
This is just to allow building a Debian source package. Since we don't
really use an upstream .tar.gz here, just tag the last commit without debian
packaging with a fake upstream tag.

It would be better to bump the upstream version as well, but that might
make a conflict if the official webrtc-audio-processing Debian package
ever decides to upload an update.